+++ src/app/journey/03-advanced-techniques/01-physics/index.ts
@@ -1,5 +1,3 @@
-import * as CANNON from 'cannon-es';
-
 import { Timer } from 'three/addons/misc/Timer.js';
 
 import { CustomizableQuality, DebugFPS } from '#/app/decorators/debug';
@@ -8,13 +6,11 @@ import OrbitControlledExercise from "#/app/journey/exercises/orbit-controlled-ex
 import RenderView from '#/app/layout/render-view';
 import { ExtraConfig } from '#/app/types/exercise';
 import { CSS_CLASSES } from '#/theme';
-import { CollisionSound } from './collision-sound';
 import BOX from './icons/cube.svg?raw';
 import SPHERE from './icons/sphere.svg?raw';
 import REMOVE from './icons/trash.svg?raw';
 import { Lighting } from './lighting';
-import { CollisionEvent, PhysicalObject } from './physical-object';
-import { PhysicalObjectFactory } from './physical-object-factory';
+import { PhysicalWorld } from './physical-world';
 import { QUALITY_CONFIG, QualityConfig } from "./quality-config";
 
 @Exercise('physics')
@@ -26,19 +22,11 @@ import { QUALITY_CONFIG, QualityConfig } from "./quality-config";
 )
 @CustomizableQuality
 export class Physics extends OrbitControlledExercise {
-  private physicalObjects: PhysicalObject[];
-
-  private floor: PhysicalObject;
-
   private lighting: Lighting;
-
-  private physicsWorld: CANNON.World;
-
-  private collisionSound: CollisionSound;
 
   private qualityConfig: QualityConfig;
 
-  private factory: PhysicalObjectFactory;
+  private physicalWorld: PhysicalWorld;
 
   constructor(view: RenderView, extraConfig: ExtraConfig) {
     super(view);
@@ -46,94 +34,40 @@ export class Physics extends OrbitControlledExercise {
     this.qualityConfig = QUALITY_CONFIG[extraConfig.quality];
     view.enableShadows(this.qualityConfig.shadowMapType);
 
-    this.physicsWorld = this.setupPhysics();
-    this.collisionSound = new CollisionSound();
-
-    this.factory = new PhysicalObjectFactory({
-      world: this.physicsWorld,
+    this.physicalWorld = new PhysicalWorld({
       scene: this.scene,
       sphereSubdivisions: this.qualityConfig.sphereSubdivisions,
-      onCollide: this.playHitSound.bind(this),
     });
-
-    this.physicalObjects = [];
-    this.floor = this.factory.createFloor();
 
     this.lighting = new Lighting();
     this.camera.position.set(-3, 3, 3);
     this.lighting.setup(this.scene);
-
   }
 
   @DebugFPS
   frame(timer: Timer) {
     super.frame(timer);
     const delta = timer.getDelta();
-    this.physicsWorld.step(1 / 60, delta, 3);
-
-    for(let i = this.physicalObjects.length - 1; i >= 0; i--) {
-      const object = this.physicalObjects[i];
-      if(object.isBelowY(-20)) {
-        this.removeObject(object, i);
-      } else {
-        object.sync();
-      }
-    }
-  }
-
-  public playHitSound(collision: CollisionEvent) {
-    const velocity = collision.contact.getImpactVelocityAlongNormal();
-    if (velocity > 1.5) {
-      this.collisionSound.play();
-    }
-  }
-
-  private setupPhysics() {
-    const world = new CANNON.World();
-    world.gravity.set(0, -9.82, 0);
-    world.broadphase = new CANNON.SAPBroadphase(world);
-    world.allowSleep = true;
-
-    const material = new CANNON.Material('default');
-    const contactMaterial = new CANNON.ContactMaterial(material, material, {
-        friction: 0.1,
-        restitution: 0.7,
-    });
-
-    world.defaultContactMaterial = contactMaterial;
-    return world;
+    this.physicalWorld.update(delta);
   }
 
   @ActionButton('Add Sphere', SPHERE)
   public addSphere() {
-    const sphere = this.factory.createSphere();
-    this.physicalObjects.push(sphere);
+    this.physicalWorld.addSphere();
   }
 
   @ActionButton('Add Box', BOX)
   public addBox() {
-    const box = this.factory.createBox();
-    this.physicalObjects.push(box);
+    this.physicalWorld.addBox();
   }
 
   @ActionButton('Remove All', REMOVE)
   public clearScene() {
-    for(let i = this.physicalObjects.length - 1; i >= 0; i--) {
-      this.removeObject(this.physicalObjects[i], i);
-    }
-  }
-
-  public removeObject(object: PhysicalObject, index: number) {
-    object.dispose();
-    this.physicalObjects.splice(index, 1);
+    this.physicalWorld.clearScene();
   }
 
   async dispose() {
     await super.dispose();
-    this.physicalObjects.forEach((object, index) => {
-      this.removeObject(object, index);
-    });
-    this.floor.dispose();
-    this.factory.dispose();
+    this.physicalWorld.dispose();
   }
 }
